uint32

+++ tmos/hardware/stm32/gpio_stm32.h
@@ -29,11 +29,11 @@ typedef unsigned int PIN_DESC;
 /// Pin definition bits 31:30 = MODE
 #define PD_MODE_Pos 		30						//!< mode bits position
 #define PD_MODE_Msk			(0x3 << PD_MODE_Pos)	//!< mode bits mask
-#define PD_MODE_Get(x)		((x) >> PD_MODE_Pos) 	//!< mode value
-#define PD_MODE_INPUT		(0 << PD_MODE_Pos)		//!< input (reset state)
-#define PD_MODE_OUTPUT		(1 << PD_MODE_Pos)		//!< general purpose output mode
-#define PD_MODE_AF			(2 << PD_MODE_Pos)		//!< alternative function mode
-#define PD_MODE_ANALOG		(3 << PD_MODE_Pos)		//!< analog mode
+#define PD_MODE_Get(x)		(uint32_t)((x) >> PD_MODE_Pos) 	//!< mode value
+#define PD_MODE_INPUT		(0u << PD_MODE_Pos)		//!< input (reset state)
+#define PD_MODE_OUTPUT		(1u << PD_MODE_Pos)		//!< general purpose output mode
+#define PD_MODE_AF			(2u << PD_MODE_Pos)		//!< alternative function mode
+#define PD_MODE_ANALOG		(3u << PD_MODE_Pos)		//!< analog mode
 
 
 /// Pin definition bits 29:29 = output type
